+++ tests/compile/derive-enum.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2020 Robin Krahl <robin.krahl@ireas.org>
 // SPDX-License-Identifier: Apache-2.0 or MIT
 
-use merge::Merge;
+use conflate::Merge;
 
 #[derive(Merge)]
 enum E {