+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 All notable changes to this project will be documented in the corresponding
 changelog files for each crate.
 
-- [merge](./crates/merge/CHANGELOG.md)
-- [merge-derive](./crates/merge_derive/CHANGELOG.md)
+- [conflate](./crates/conflate/CHANGELOG.md)
+- [conflate-derive](./crates/conflate_derive/CHANGELOG.md)